BONUS: 15min passed since payment calculatePrice() price logic
Before this commit calculatePrice() always took the date of the issuance of the ticket as the reference date in regards to the price of the ticket. This was not correct, because if a ticket counts as unpaid due to 15 min have passed since the payment, then the driver should not have to pay the full time again, but only the time since the last payment.

Furthermore there were some small adjustments within calculateTicketState() in regards to exactly 15 min do not count as UNPAID anymore.